fix spring circular dependencies
