+++ src/main/java/eu/europa/ec/dgc/issuance/config/HcertLibConfig.java
@@ -16,7 +16,6 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 @RequiredArgsConstructor
 public class HcertLibConfig {
-
     private final EhdCryptoService ehdCryptoService;
 
     @Bean
